# packages/spike.land/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Spike Land API
  description: |
    API for Spike Land - AI-Powered App Platform with Image Enhancement

    ## Platform Overview
    Spike Land is a platform where users create, fork, and manage AI-powered applications.
    The flagship application, Pixel, provides AI-powered image enhancement.

    ## Authentication
    All protected endpoints require a valid session cookie from NextAuth.js.
    Alternatively, you can use Bearer token authentication for API requests:
    ```
    Authorization: Bearer {session_token}
    ```

    ## Token Economy
    Users have a platform-wide token balance used for:
    - Image enhancement operations (cost depends on resolution tier)
    - Premium features in apps

    Token costs for Pixel image enhancement:
    - TIER_1K: 2 tokens
    - TIER_2K: 5 tokens
    - TIER_4K: 10 tokens

    Tokens regenerate automatically (1 token per 15 minutes, max 100).

    ## Rate Limiting
    Most endpoints are rate-limited per user. Rate limit information is included in response headers:
    - X-RateLimit-Remaining: Requests remaining in current window
    - X-RateLimit-Reset: Unix timestamp when limit resets

    Example rate limits:
    - Image enhancement: 3 requests per minute
    - Image upload: 10 requests per minute
    - Album operations: 10 requests per minute

    ## Error Handling
    The API uses standard HTTP status codes:
    - 200 OK: Successful request
    - 201 Created: Resource created
...

...

    ## Base URL
...
servers:
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (automatically set after login)
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for API requests
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Unauthorized"
        code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
        details:
          type: object
          description: Additional error context
          example: { "field": "email", "reason": "invalid format" }
        title:
          type: string
          description: Short error title
          example: "Authentication Failed"
        suggestion:
          type: string
          description: Suggested action to resolve the error
          example: "Please log in again"
    ValidationError:
      type: object
      required:
        - error
        - field
      properties:
        error:
          type: string
          description: Validation error message
          example: "Invalid input"
        field:
          type: string
          description: Field that failed validation
          example: "email"
        code:
          type: string
          enum:
            - REQUIRED_FIELD
            - INVALID_FORMAT
            - FILE_TOO_LARGE
            - UNSUPPORTED_FORMAT
          description: Validation error code
          example: "INVALID_FORMAT"

    # ========================================
    # Enumerations
    # ========================================
    EnhancementTier:
      type: string
      enum:
        - TIER_1K
        - TIER_2K
        - TIER_4K
      description: Image enhancement resolution tier
      example: TIER_2K
    JobStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED
        - REFUNDED
        - CANCELLED
      description: Status of an enhancement job
      example: PROCESSING
...
    Image:
      type: object
      required:
        - id
        - userId
        - originalUrl
        - name
        - mimeType
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Image ID (CUID)
          example: "cuid123456789"
        userId:
          type: string
          description: Owner user ID
          example: "cuid987654321"
        originalUrl:
          type: string
          format: uri
          description: URL to original uploaded image
          example: "https://storage.example.com/original_abc123.jpg"
        name:
          type: string
          description: Display name of the image
          example: "My Photo"
        mimeType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          description: Image MIME type
          example: "image/jpeg"
        albumId:
          type: string
          description: Album ID if image belongs to an album
          example: "cuid456789123"
        createdAt:
          type: string
          format: date-time
          description: Image creation timestamp
          example: "2025-12-10T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-12-12T15:45:00Z"
...
    ImageEnhancementJob:
      type: object
      required:
        - id
        - userId
        - imageId
        - tier
        - status
        - tokensCost
        - createdAt
      properties:
        id:
          type: string
          description: Job ID (CUID)
          example: "job_abc123"
        userId:
          type: string
          description: User ID
          example: "cuid987654321"
        imageId:
          type: string
          description: Image ID being enhanced
          example: "cuid123456789"
        tier:
          $ref: "#/components/schemas/EnhancementTier"
        status:
          $ref: "#/components/schemas/JobStatus"
        tokensCost:
          type: integer
          description: Tokens consumed by this job
          example: 5
        enhancedImageId:
          type: string
          description: ID of resulting enhanced image (when completed)
          example: "cuid_enh_123"
        error:
          type: string
          description: Error message if job failed
          example: "Image processing timeout"
        progress:
          type: integer
          description: Job progress percentage (0-100)
          example: 75
        createdAt:
          type: string
          format: date-time
          description: Job creation timestamp
          example: "2025-12-12T14:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-12-12T14:35:00Z"
        completedAt:
          type: string
          format: date-time
          description: Job completion timestamp
          example: "2025-12-12T14:40:00Z"

    # ========================================
    # Album Models
    # ========================================
...
  responses:
    Unauthorized:
      description: Authentication required or failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
            title: "Authentication Failed"
            suggestion: "Please log in again"
    Forbidden:
      description: User lacks permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
            title: "Access Denied"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
            title: "Not Found"
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/Error"
              - $ref: "#/components/schemas/ValidationError"
          example:
            error: "Invalid input"
            code: "INVALID_INPUT"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp when limit resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Rate limit exceeded"
            code: "RATE_LIMIT_EXCEEDED"
            title: "Too Many Requests"
            suggestion: "Please wait before making another request"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Internal server error"
            code: "INTERNAL_SERVER_ERROR"
            title: "Server Error"
            suggestion: "Please try again later"

  # ========================================
  # Request Parameters
  # ========================================
  parameters:
...
    IdParam:
      name: id
      in: path
      required: true
      description: Resource ID (CUID)
      schema:
        type: string
      example: "cuid123456789"
...
paths:
...
  /api/images/upload:
    post:
      summary: Upload a single image
      tags: [Images]
      operationId: uploadImage
      description: |
        Upload a single image to the platform. Supported formats: JPEG, PNG, WebP, HEIC.
        Maximum file size: 10MB. Image metadata is extracted and stored for enhancement tracking.
        Optional albumId can be provided to add image directly to an album upon upload.
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP, HEIC)
                albumId:
                  type: string
                  description: Optional album ID to add image to upon upload
      responses:
        "200":
          description: Image uploaded successfully
          headers:
            X-Request-ID:
              schema:
                type: string
              description: Unique request identifier for troubleshooting
          content:
            application/json:
              schema:
                type: object
                required: [success, image]
                properties:
                  success:
                    type: boolean
                  image:
                    $ref: "#/components/schemas/Image"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/images/{id}:
    get:
      summary: Get image details with enhancement jobs
      tags: [Images]
      operationId: getImage
      description: |
        Retrieve detailed information about a specific image including all associated
        enhancement jobs. Returns full image metadata and list of all enhancement jobs
        with their status and results. User must be the image owner or image must be public.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Image details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required: [success, image]
                properties:
                  success:
                    type: boolean
                  image:
                    allOf:
                      - $ref: "#/components/schemas/Image"
                      - type: object
                        properties:
                          jobs:
                            type: array
                            items:
                              $ref: "#/components/schemas/ImageEnhancementJob"
                            description: List of enhancement jobs for this image
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      summary: Delete an image
      tags: [Images]
      operationId: deleteImage
      description: |
        Delete an image and all associated enhancement jobs. Also removes corresponding
        files from R2 storage. User must be the image owner. Operation is transactional
        - R2 files are deleted first, then database records removed.
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Image deleted successfully
          content:
            application/json:
              schema:
                type: object
                required: [success, message]
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"